# Perf tests: ReactNative Bazel C++ example