--- a/sample/Frankenstein/Makefile
+++ b/sample/Frankenstein/Makefile
@@ -1,0 +1,3 @@
+deps:
+	git clone https://github.com/pinterest/PodToBUILD.git Vendor/rules_pods
+	Vendor/rules_pods/bin/update_pods.py