--- a/sample/Frankenstein/WORKSPACE
+++ b/sample/Frankenstein/WORKSPACE
@@ -1,0 +1,20 @@
+load("@bazel_tools//tools/build_defs/repo:git.bzl", system_git_repository = "git_repository")
+
+git_repository(
+    name = "build_bazel_rules_apple",
+    commit = "6f93226e29fef0ac6ba00fb5dc56389d3bd8c25f",
+    remote = "https://github.com/bazelbuild/rules_apple.git",
+)
+
+load(
+    "@build_bazel_rules_apple//apple:repositories.bzl",
+    "apple_rules_dependencies",
+)
+
+apple_rules_dependencies()
+
+http_file(
+    name = "xctestrunner",
+    executable = 1,
+    urls = ["https://github.com/google/xctestrunner/releases/download/0.2.3/ios_test_runner.par"],
+)