--- a/sample/Frankenstein/XCHammer.yaml
+++ b/sample/Frankenstein/XCHammer.yaml
@@ -1,0 +1,7 @@
+targets:
+    - ":ios-app"
+
+projects:
+    "Frankenstein":
+        paths:
+            - "**"