--- a/sample/Frankenstein/main.m
+++ b/sample/Frankenstein/main.m
@@ -1,0 +1,3 @@
+int main(int argc, char * argv[]) {
+    return 0;
+}