… sample/Frankenstein/Vendor/PodSpecs/boost-for-react-native-1.63.0-0/boost-for-react-native.podspec
… sample/Frankenstein/Vendor/PodSpecs/glog-0.3.4/ios-configure-glog.sh
… sample/Frankenstein/Vendor/PodSpecs/react-0.57/third-party-podspecs/DoubleConversion.podspec
… sample/Frankenstein/Vendor/PodSpecs/react-0.57/third-party-podspecs/Folly.podspec
… sample/Frankenstein/Vendor/PodSpecs/react-0.57/third-party-podspecs/glog.podspec