This commit adds another example of a project containing RN and all of
it's trans deps.

On the CI, we run XCHammer to just build the Xcode project as a perf
test of this condition.

Now that PodToBUILD can build ReactNative, this is easy to setup. Future
work will clean this up when we add support for external in P2B

On a maxed out 2.9 GHz Intel Core i7 MBP
```
XCHammer: ** Completed xcode_gen in 378.3899s
XCHammer: ** Completed make_schemes in 0.0283s
XCHammer: ** Completed generate_project in 380.5344s

```
This is a highly optimized case in XCHammer logic ( only 2s here ) and
exemplifies the XcodeGen issue