@jerrymarino 2018-10-08
I don't think we can rely on Travis CI for exact perf benchmarks, but this should give us a rough approximation of what we can expect.

The last run clearly captures the issue: 

```
XCHammer: ** Completed convert_targets in 3.6889s
XCHammer: Writing project
XCHammer: ** Completed xcode_gen in 428.4386s
XCHammer: ** Completed make_schemes in 0.0321s
XCHammer: ** Completed generate_project in 432.1937s
XCHammer: Replacing old project
XCHammer: Overwrote project```